add delete method for controller v3

+++ src/main/groovy/com/br/fatec/edu/todoapp/todo/service/TodoServiceImpl.groovy
@@ -2,7 +2,10 @@ package com.br.fatec.edu.todoapp.todo.service
 
 import com.br.fatec.edu.todoapp.todo.model.Todo
 import com.br.fatec.edu.todoapp.todo.repository.TodoRepository
+
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 
 @Service
@@ -12,13 +15,15 @@ class TodoServiceImpl implements TodoService {
     TodoRepository todoRepository
     
     @Override
-    List<Todo> findAll() {
-        todoRepository.findAll()
+    Page<Todo> findAll(Pageable pageable) {
+        return todoRepository.findAll(pageable)
     }
 
     @Override
     Todo findById(Integer todoId) {
-        todoRepository.findById(todoId).get()
+        if(todoRepository.existsById(todoId)) {
+            return todoRepository.findById(todoId).get()
+        }
     }
 
     @Override
@@ -28,6 +33,8 @@ class TodoServiceImpl implements TodoService {
 
     @Override
     void deleteTodoById(Integer todoId) {
-        todoRepository.deleteById(todoId)
+        if(todoRepository.existsById(todoId)) {
+            todoRepository.deleteById(todoId)
+        }
     }
 }
